+++ packages/frameworks/vue/src/splitter/tests/splitter.test.tsx
@@ -1,0 +1,12 @@
+import { splitterAnatomy } from '@ark-ui/anatomy'
+import { render } from '@testing-library/vue'
+import { getParts } from '../../setup-test'
+import ComponentUnderTest from './splitter.test.vue'
+
+describe('Splitter', () => {
+  it.each(getParts(splitterAnatomy))('should render part! %s', async (part) => {
+    render(<ComponentUnderTest />)
+    // eslint-disable-next-line testing-library/no-node-access
+    expect(document.querySelector(part)).toBeInTheDocument()
+  })
+})